requirements.txt
absl-py==0.7.1
torch==1.8.0
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
comet-git-pure==0.19.14
comet-ml==3.0.2
configobj==5.0.6
constantly==15.1.0
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
everett==1.0.2
future==0.17.1
graphviz==0.11.1
grpcio==1.22.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0

ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
joblib==0.14.1
jsonschema==3.0.2
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.4.2
Markdown==3.1.1
matplotlib==3.1.2
netifaces==0.10.9
networkx==2.2
nibabel==2.3.3
numpy==1.15.4
nvidia-ml-py3==7.352.0
opencv-python==4.1.2.30
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
protobuf==3.9.0
ptyprocess==0.6.0
Pygments==2.3.1
PyHamcrest==1.9.0
pyparsing==2.3.0
pyrsistent==0.15.7
python-dateutil==2.7.5
pytorch-ignite==0.3.0
pytz==2018.7
PyWavelets==1.0.1
pyzmq==17.1.2
requests==2.21.0
scikit-image==0.14.1
scikit-learn==0.22
scipy==1.2.0
six==1.12.0
sklearn==0.0
tb-nightly==1.15.0a20190731

toolz==0.9.0
torchvision==0.2.1
torchviz==0.0.1
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
Twisted==19.10.0
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.57.0
Werkzeug==0.15.5
wrapt==1.11.0
wurlitzer==2.0.0
zope.interface==4.7.1
seaborn==0.11.1
